Separate `KeyValueStoreTest` logic
